message_edit: Restrict editing/deleting in deactivated streams.
Users were able to edit/delete their messages in deactivated streams.
Added checks to see if a message is part of a deactivated stream
before allowing these actions by non admin users.
If a user attempts to edit or delete a message in a deactivated
stream an error is now displayed to the user.
Added backend testing for associated changes.

Fixes #17231.